+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project should be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## 2018-10-31 - Minor Refactoring
+
+### Added
+
+- Added Season class to manage seasonality
+
+### Changed
+
+- Removed resolution from the Raster class
+
 ## 2018-09-20 - Variable Rename
 
 ### Added